+++ CHANGELOG.md
@@ -103,6 +103,9 @@ Miscellaneous:
 - 3D viewer: the accurateness of object picking on neurons and location picking
   on z panes has been improved.
 
+- Cropping image data works again.
+
+
 ## 2017.05.17
 
 Contributors: Andrew Champion, Tom Kazimiers